Implement user-settable video speeds
Adds two preferences where users can specify the video speeds and and video speed modifiers
that will be displayed. The preferences are comma-separated numbers. To support this,
a separate module was created so that it could be tested to see if the numbers were being
parsed correctly.

Update build and readme, bump version