Show menu even if media is paused (#4)
We used to only show the menu if there was a single un-paused media
element. We did this because it was easier to identify which element
should be manipulated. We now first check to see if there is only one
media element anyway and just act on that. If there is more than one, we
use not being paused as the criterion to find the element. Bump version
and update build